requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
browser-cookie3==0.13.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
genshinstats==1.4.11
idna==3.3
importlib-metadata==4.8.1
jeepney==0.7.1
keyring==23.2.1
lz4==3.1.3
pbkdf2==1.3
pyaes==1.6.1
pycparser==2.20
pycryptodome==3.11.0
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
SecretStorage==3.3.1
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
zipp==3.6.0
